fix(combobox): update docs examples
