+++ packages/paste-website/src/pages/components/combobox/index.mdx
@@ -200,7 +200,7 @@ It should be noted that when doing so, the `state` prop takes precident over the
 
 For full details on how to use the state hook, and what props to provide it, follow the [Combobox Primitive documentation](/primitives/combobox-primitive#usecomboboxprimitive-arguments). It's the same hook, just renamed.
 
-<LivePreview scope={{Combobox, Box, Button, CloseIcon, SearchIcon, useCombobox}} noInline language="jsx">
+<LivePreview scope={{Button, Box, CloseIcon, SearchIcon, Combobox, useCombobox}} noInline language="jsx">
   {stateHookCombobox}
 </LivePreview>
